back/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
EbookLib==0.19
lxml==6.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
